+++ src/misc/providers.ts
@@ -1,7 +1,9 @@
-import {ClassProvider, EnvironmentProviders, Type, ValueProvider, makeEnvironmentProviders} from '@angular/core';
+import {ClassProvider, EnvironmentProviders, FactoryProvider, Provider, Type, ValueProvider, makeEnvironmentProviders, inject} from '@angular/core';
+import {extend} from '@jscrpt/common';
 
-import {ANGLR_EXCEPTION_EXTENDERS, INTERNAL_SERVER_ERROR_RENDERER} from './tokens';
+import {ANGLR_EXCEPTION_EXTENDERS, HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_MESSAGES, HTTP_CLIENT_ERROR_RESPONSE_MAPPER, HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER, INTERNAL_SERVER_ERROR_RENDERER} from './tokens';
 import {AnglrExceptionExtender, InternalServerErrorRenderer} from '../interfaces';
+import {HttpClientErrorHandlers, HttpClientErrorMessages, HttpClientErrorResponseMapper, HttpClientValidationErrorResponseMapper} from './types';
 
 /**
  * Provides anglr exception extends functions
@@ -33,4 +35,62 @@ export function provideInternalServerErrorRenderer(renderer: Type<InternalServer
             useClass: renderer,
         }
     ]);
+}
+
+/**
+ * Provides client error messages and merges them with existing provided messages
+ * @param messages - Object storing messages to be provided
+ */
+export function provideClientErrorMessages(messages: HttpClientErrorMessages): Provider
+{
+    return <FactoryProvider> {
+        provide: HTTP_CLIENT_ERROR_MESSAGES,
+        useFactory: () =>
+        {
+            const existingMessages = inject(HTTP_CLIENT_ERROR_MESSAGES, {skipSelf: true, optional: true});
+
+            return extend({}, existingMessages, messages);
+        },
+    };
+}
+
+/**
+ * Provides client error handlers and merges them with existing provided handlers
+ * @param handlers - Object storing handlers to be provided
+ */
+export function provideClientErrorHandlers(handlers: HttpClientErrorHandlers): Provider
+{
+    return <FactoryProvider> {
+        provide: HTTP_CLIENT_ERROR_HANDLERS,
+        useFactory: () =>
+        {
+            const existingHandlers = inject(HTTP_CLIENT_ERROR_HANDLERS, {skipSelf: true, optional: true});
+
+            return extend({}, existingHandlers, handlers);
+        },
+    };
+}
+
+/**
+ * Provides http client error response mapper function
+ * @param mapper - Mapper function to be provided
+ */
+export function provideHttpClientErrorResponseMapper(mapper: HttpClientErrorResponseMapper): Provider
+{
+    return <ValueProvider> {
+        provide: HTTP_CLIENT_ERROR_RESPONSE_MAPPER,
+        useValue: mapper,
+    };
+}
+
+/**
+ * Provides http client validation error response mapper function
+ * @param mapper - Mapper function to be provided
+ */
+export function provideHttpClientValidationErrorResponseMapper(mapper: HttpClientValidationErrorResponseMapper): Provider
+{
+    return <ValueProvider> {
+        provide: HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER,
+        useValue: mapper,
+    };
 }